Don't convert quotes when Vim is not inserting
The Vim plugin uses the { key as a motion key, except in insert mode. This change makes this extension only convert quotes when Vim is not in an insert mode. Since the variable `vim.mode` is not defined when the Vim extension is not installed, a very long list of modes that should suppress this extension is used. That way quote replacing behavior works both with and without the Vim extension installed.